Sets rating on both tablet and phone

+++ app/src/main/java/com/example/android/historyapp/ui/SeeSitesActivity.java
@@ -49,7 +49,7 @@ public class SeeSitesActivity extends AppCompatActivity implements SitesAdapter.
             arguments.putString(SitesDetailsFragment.ARG_ITEM_ID, SitesListFragment.dummyD.get(i).getName());
             arguments.putString(SitesDetailsFragment.ARG_ITEM_HISTORY, SitesListFragment.dummyD.get(i).getHistory());
             arguments.putString(SitesDetailsFragment.ARG_ITEM_OPERATOR, SitesListFragment.dummyD.get(i).getOperator());
-            arguments.putDouble(SitesDetailsFragment.ARG_ITEM_RATING, SitesListFragment.dummyD.get(i).getRating());
+            arguments.putInt(SitesDetailsFragment.ARG_ITEM_RATING, SitesListFragment.dummyD.get(i).getRating());
             SitesDetailsFragment fragment = new SitesDetailsFragment();
             fragment.setArguments(arguments);
